+++ shared/app/app/static/src/main.js
@@ -1,6 +1,6 @@
 // @flow
 
-import * as tools from './tools';
+import * as tools from './tools/tools';
 
 tools.Console(tools.add(2, 10).toString());
 tools.Console(tools.mulToString(2, 5).toString());
@@ -8,5 +8,5 @@ tools.Console(tools.mulToString(2, 5).toString());
 const myId:HTMLElement | null = document.getElementById('intro');
 
 if (myId instanceof HTMLElement) {
-  myId.innerHTML = '<p>Hello from JS</p>';
+  myId.innerHTML = '<p>Hello from JS!</p>';
 }
