+++ shared/app/app/static/src/tools/tools.test.js
